Reduce assumptions about locale's behavior in new regex tests.
I was overoptimistic to assume that UTF8-based locales would all
consider U+1500 to be a member of the [[:alpha:]] char class.
Tweak the test cases added by commit a44cdf636 to avoid that
assumption.  We might need to lobotomize them further, but this
should be enough to fix the early buildfarm reports.